[메카인] #MMOFighter, Character의 x, y 값을 int로 고정하고 패킷 생성시에만 형 변환 하도록 수정

+++ Practice/MMOFighterServerChoi/MMOFighterServerChoi/Character.h
@@ -9,8 +9,8 @@ struct Character
 
     BYTE action = 0;
     BYTE direction = 0;
-    short x = 0;
-    short y = 0;
+    int x = 0;
+    int y = 0;
     int HP = 0;
 
     SectorPos sector;
